requirements.txt
Pillow==11.0.0
pillow-heif==0.21.0